refactor: BrowserRouter 대신 createBrowserRouter로 변경
